refactor: 주문 관리 페이지 상단 메뉴바 transform 수정
OptionContainer
- props를 div 속성을 받도록 수정하였습니다.
- styles를 분리하였습니다.
- 검색창에서 빈 값을 입력시에 query string을 제거하도록 searchOrder를 변경하였습니다.
useStikyHeader의 이름을 useElementTopOffset으로 보다 명확하게 변경하였습니다.
ManageOrderPage
- 주문 관리 메뉴바가 일정 높이 이상 지날 경우 상단 fixed 메뉴로 변경되도록 값을 변경하였습니다.
- handleMotionDiff는 useTransform에서 사용될 배열을 리턴하는 함수입니다.
