+++ src/Page/MangeOrder/MangeOrderPage.tsx
@@ -4,19 +4,46 @@ import {
   OrderStateListContainer
 } from "@/Components";
 import { Container, OrderStateContainer, OrderStateHeader } from "./styles";
-import { useGetOrder, useStikyHeader } from "./hooks";
+import { useGetOrder, useElementTopOffset } from "./hooks";
+import { useTransform, useViewportScroll } from "framer-motion";
+
+const handleMotionDiff = <TStart, TDiff>(start: TStart, ...rest: TDiff[]) => {
+  return [start, ...rest];
+};
 
 const ManageOrderPage = () => {
   const { isLoading, isRefetching } = useGetOrder();
-  const { sticky, stickyPos } = useStikyHeader();
+  const { elemetRef, topOffset } = useElementTopOffset();
+
+  const { scrollY } = useViewportScroll();
+
+  const opacity = useTransform(
+    scrollY,
+    handleMotionDiff(0, topOffset, topOffset + 1),
+    handleMotionDiff(0, 0, 1)
+  );
+  const borderTick = useTransform(
+    scrollY,
+    handleMotionDiff(0, topOffset, topOffset + 1),
+    handleMotionDiff("unset", "unset", "3px solid #575757")
+  );
 
   return isLoading && isRefetching ? (
     <Loading title="주문 정보를 불러오고있어요" />
   ) : (
     <Container>
-      <OrderStateHeader ref={sticky}>
-        <OptionsContainer stickyPos={stickyPos} />
+      <OrderStateHeader
+        style={{
+          opacity,
+          borderBottom: borderTick,
+          zIndex: 100
+        }}
+      >
+        <OptionsContainer />
       </OrderStateHeader>
+      <div ref={elemetRef}>
+        <OptionsContainer />
+      </div>
       <OrderStateContainer>
         <OrderStateListContainer />
       </OrderStateContainer>
