+++ src/Page/MangeOrder/hooks/index.ts
@@ -1,2 +1,2 @@
 export { default as useGetOrder } from "./useGetOrder";
-export { default as useStikyHeader } from "./useStikyHeader";
+export { default as useElementTopOffset } from "./useElementTopOffset";
