+++ src/Page/MangeOrder/styles.ts
@@ -1,5 +1,12 @@
+import { motion } from "framer-motion";
 import styled from "styled-components";
 
 export const Container = styled.div``;
-export const OrderStateHeader = styled.div``;
+export const OrderStateHeader = styled(motion.div)`
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  padding: 0 2rem;
+`;
 export const OrderStateContainer = styled.div``;
